+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.4.0] - 2020-05-26
+### Changed
+- Add support for adding and using tokenized cards as payment methods
+- Add tokenized card CIT payments functionalities
+- Add Callback controller class and implementation
+
 ## [1.3.1] - 2020-04-15
 ### Fixed
 - Fixed the hosted payment method selection